server/requirements.txt
python-dotenv
pydantic
uvicorn[standard]
gunicorn
fastapi
fastapi-utils
typing
pymongo
yapf
pylint
